restructure the property types API to make tiled.project.propertyTypes a list of types and move the utility methods for custom types to tiled.project
