+++ src/tiled/editableproject.h
@@ -23,7 +23,7 @@
 
 #include "editableasset.h"
 #include "project.h"
-#include "scriptpropertytypes.h"
+#include "scriptpropertytype.h"
 
 #include <QObject>
 
@@ -39,8 +39,7 @@ class EditableProject final : public EditableAsset
     Q_PROPERTY(QString automappingRulesFile READ automappingRulesFile)
     Q_PROPERTY(QString fileName READ fileName)
     Q_PROPERTY(QStringList folders READ folders)
-    Q_PROPERTY(ScriptPropertyTypes *propertyTypes READ propertyTypes)
-
+    Q_PROPERTY(QVector<ScriptPropertyType *> propertyTypes READ propertyTypes)
 public:
     EditableProject(ProjectDocument *projectDocument, QObject *parent = nullptr);
 
@@ -51,10 +50,16 @@ public:
     QString automappingRulesFile() const;
     QString fileName() const;
     QStringList folders() const;
-    ScriptPropertyTypes *propertyTypes() const;
     Project *project() const;
 
     QSharedPointer<Document> createDocument() override;
+
+    Q_INVOKABLE void removeTypeByName(const QString &name);
+    Q_INVOKABLE ScriptPropertyType *findTypeByName(const QString &name);
+    QVector<ScriptPropertyType*> propertyTypes() const;
+private:
+    ScriptPropertyType *toScriptType(const PropertyType *type) const;
+    void applyPropertyChanges();
 };
 
 inline Project *EditableProject::project() const
