+++ src/tiled/libtilededitor.qbs
@@ -458,8 +458,8 @@ DynamicLibrary {
         "scriptmodule.h",
         "scriptprocess.cpp",
         "scriptprocess.h",
-        "scriptpropertytypes.cpp",
-        "scriptpropertytypes.h",
+        "scriptpropertytype.cpp",
+        "scriptpropertytype.h",
         "selectionrectangle.cpp",
         "selectionrectangle.h",
         "selectsametiletool.cpp",
